Jeff cleanup2025

+++ app/src/main/java/com/team3663/scouting_app/activities/QRCode.java
@@ -272,7 +272,7 @@ public class QRCode extends AppCompatActivity {
 
             while (begin < size) {
                 if (size - begin <= Globals.CurrentQRSize) end = size;
-                else end = begin + (int)(Globals.CurrentQRSize * Constants.QRCode.QR_PREFERRED_SIZE_PERCENTAGE / 100);
+                else end = begin + (Globals.CurrentQRSize * Constants.QRCode.QR_PREFERRED_SIZE_PERCENTAGE / 100);
 
                 file_page.add(in_data.substring(begin, end));
                 begin = end;
